Expose all data constructors via .Internal modules (close #305)
